==> data-prep/02_ais_data.Rmd <==
---
title: "Vessel Tracks"
output: html_document
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(
	echo = FALSE,
	message = FALSE,
	warning = FALSE
)
```

```{r libraries, include=F}
# Load libraries
library(tidyverse)
library(glue)
library(stringr)
library(sf)
library(paletteer)
library(grDevices)
library(colorspace)
library(bigrquery)
library(DBI)

# Load source
source(file.path(here::here(),"common.R"))
```

# Bounding Box 

```{r check-bbox}
# Central CA coastline
ca_counties <- tigris::counties() %>% 
  filter(GEOID %in% c('06083', '06111', '06037'))

# CI polygons
ci_polys <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands.shp"))

# bbox coords xmin: -121, xmax: -118; ymin: 32.5, ymax: 34.5
bbox_coords <- matrix(c(-121, 32.5,
                        -118, 32.5,
                        -118, 34.5,
                        -121, 34.5,
                        -121, 32.5),
                        ncol = 2, byrow = TRUE)

bbox_poly <-st_polygon(list(bbox_coords))
bbox_poly <- st_sfc(bbox_poly)
st_crs(bbox_poly) <- 4326

ggplot() + 
  geom_sf(data = ca_counties) +
  geom_sf(data = ci_polys) +
  geom_sf(data = bbox_poly, color = 'midnightblue',
          fill = NA)
```

# Query AIS Data

Query all AIS data for all ssvids (15) on the AIS vessel list. 

```{r ais-data, eval=F}
# Run in BQ console b/c of the create or replace statement
# Query is very large 1.3Tb 
# Saved as emlab-gcp.boi_dive_project.ais_activity_all_years
sql_ais <- "#StandardSQL

CREATE OR REPLACE TABLE
  emlab-gcp.boi_dive_project.ais_activity_all_years
  PARTITION BY date
  CLUSTER BY ssvid
  AS

WITH 

good_segments AS (
  SELECT
    seg_id
  FROM
    `world-fishing-827.pipe_production_v20201001.research_segs`
  WHERE good_seg
  AND NOT overlapping_and_short
),

ais AS (
 SELECT
   *,
   EXTRACT(DATE FROM timestamp) AS date
 FROM
  `world-fishing-827.pipe_production_v20201001.research_messages`
 WHERE
   seg_id IN (
     SELECT
       seg_id
      FROM
        good_segments
   )
  AND ssvid IN ('338350134',
  '368200240',
  '368926530',
  '367166830',
  '366813530',
  '367036830',
  '367130890',
  '367315630',
  '338776000',
  '367616930',
  '367655170',
  '368024590',
  '367041160',
  '367098480',
  '367384840'
  )
)

SELECT * FROM ais"
```

**Updated December 2022**  

Add AIS until November 30, 2022. Previous query pulled data through July 24, 2022. 

```{r add-2022-ais-data, eval=F}
sql_ais_2022 <- "#StandardSQL

WITH 

good_segments AS (
  SELECT
    seg_id
  FROM
    `world-fishing-827.pipe_production_v20201001.research_segs`
  WHERE good_seg
  AND NOT overlapping_and_short
),

ais AS (
 SELECT
   *,
   EXTRACT(DATE FROM timestamp) AS date
 FROM
  `world-fishing-827.pipe_production_v20201001.research_messages`
 WHERE
 _PARTITIONTIME BETWEEN '2022-07-25' AND '2022-11-30'
 AND seg_id IN (
     SELECT
       seg_id
      FROM
        good_segments
   )
  AND ssvid IN ('338350134',
  '368200240',
  '368926530',
  '367166830',
  '366813530',
  '367036830',
  '367130890',
  '367315630',
  '338776000',
  '367616930',
  '367655170',
  '368024590',
  '367041160',
  '367098480',
  '367384840'
  )
)

SELECT * FROM ais"
```

```{r append-new-data}
# Append new data to the ais_activity_all_years table
bq_project_query("emlab-gcp", 
                 sql_ais_2022,
                 destination_table = bq_table(project = "emlab-gcp",
                                              dataset = "boi_dive_project",
                                              table = "ais_activity_all_years"),
                 allowLargeResults = TRUE,
                 write_disposition = "WRITE_APPEND")
```

Create a second table of just the activity within the Channel Islands bounding box for easier processing  

Bounding box: 
  xmin: -121, xmax: -118
  ymin: 32.5, ymax: 34.5
  
```{r}
sql_ci <- "#StandardSQL

SELECT
  *
FROM
  `emlab-gcp.boi_dive_project.ais_activity_all_years`
WHERE
  (lon >= -121 AND lon <= -118)
  AND (lat >= 32.5 AND lat <= 34.5)"

# Run query and write results to BigQuery
bq_project_query("emlab-gcp",
                 sql_ci,
                 destination_table = bq_table(project = "emlab-gcp",
                                              table = "ci_ais_activity_all_years",
                                              dataset = "boi_dive_project"),
                 use_legacy_sql = FALSE, allowLargeResults = TRUE,
                 write_disposition = 'WRITE_TRUNCATE')
```
  
## Visualize AIS Data

Visualize where vessels are moving throughout the Channel Islands using 2021 as a test year.    

### Tracks

```{r}
# Run and download 
tracks <- bq_table_download('emlab-gcp.boi_dive_project.ci_ais_activity_2021',
                            n_max = Inf) %>% 
  dplyr::select(ssvid, timestamp, lat, lon, hours, date)

# Save 
write_csv(tracks, file.path(project_data_path, "processed", "ais_tracks_2021.csv"))
```

```{r tracks}
ais_list <- read_csv(file.path(project_data_path, "processed", "ais_vessel_list.csv"))

# Read-in tracks
tracks <- read_csv(file.path(project_data_path, "processed", "ais_tracks_2021.csv")) %>% 
  mutate(ssvid = as.character(ssvid))

# Unique ssvids
track_ssvids <- unique(tracks$ssvid)

# Create labels for legend 
pal_labels <- ais_list %>% 
  filter(ssvid %in% track_ssvids) %>% 
  mutate(label = paste0(ssvid, ": ", str_to_sentence(n_shipname))) %>% 
  arrange(label) %>% 
  .$label

# Colors: Everglades pal + purple, grey, orange, light green from Redwoods + dark blue from BlueRidgePkwy (National park colors) 
track_pal <- c('#91D5DEFF', '#2E8289FF', '#B4674EFF', '#EAAE37FF', '#565F41FF', 
               '#6E687EFF', "#F17236FF", "#BDB2A7FF", "#769370FF", "#163343FF")

# Plot vessel tracks (all)
track_plot <- ggplot() + 
  geom_sf(data = ca_counties) + 
  geom_sf(data = ci_polys) + 
  geom_point(data = tracks,
             aes(x = lon, 
                 y = lat,
                 color = ssvid),
             alpha = 0.4) + 
  geom_path(data = tracks,
            aes(x = lon,
                y = lat,
                color = ssvid),
            alpha = 0.8) +
  scale_color_manual(values = track_pal,
                     labels = pal_labels) + 
  labs(title = "Dive Boat Vessel Tracks",
       subtitle = "January 1, 2021 - December 31, 2021",
       color = "") + 
  theme_bw() + 
  theme(axis.line = element_blank(), 
        axis.text = element_blank(),
        axis.ticks = element_blank(), 
        axis.title = element_blank(),
        panel.background = element_blank())
```

```{r save-figure}
ggsave(plot = track_plot,
       filename = file.path(project_figure_path, "dive_vessel_track_map_2021.png"),
       dpi = 300)
```
