==> data-prep/04b_explore_dives.Rmd <==
---
title: "Explore Dives"
output: 
  bookdown::html_document2:
    toc: true
    toc_float: true
    number_sections: false
    self_contained: true
---

```{r setup, include=F, message=F, warning=F}
library(tidyverse)
library(sf)
library(paletteer)

# Load source
source(file.path(here::here(),"common.R"))

# Shapefiles
ci_shp <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands.shp"))
mpa_shp <- st_read(file.path(project_data_path, "processed", "spatial", "channel_islands_mpas.shp"))

# Data 
dive_sites <- read_csv(file.path(project_data_path, "processed", "dive_sites_1hr30min_1vessel_labeled.csv"))
ais_dives  <- read_csv(file.path(project_data_path, "processed", "ais-dives", "subset_north_max5hr_daytime.csv"))
```

## Interactive Map

Map of the Northern Channel Island MPAs and possible dive sites. The map includes the following layers:  

  - `mpa_outline`: outline of MPA boundaries  
  - `mpa_type`: type of MPA (e.g., Marine Reserve or Conservation Area)   
  - `dive_sites`: possible dive sites identified through AIS (e.g., all sites where a vessel remained stationary for at least 90 minutes)  
    - There are 1,526 possible dive sites identified in the Northern Channel Islands  
  - `ais_dives`: subset of dives identified from AIS tracks from January 1, 2016 - November 30, 2022 
    - There are 3,537 dives across 10 dive vessels; these are only dives in Northern Channel Islands, where vessels remained at site between 1.5 and 5 hours and when dives occurred during the day (i.e., start time between 6 am and 6 pm).     
  
Click the layers in the upper right hand corner to add layers to the map. Hover over MPAs in the `mpa_type` layer for MPA names, and over the dots in the `ais_dives` layer for more information about that dive (e.g., shipname, date, time of day). Both the `dive_sites` and `ais_dives` layers are clustered, zooming in on the map with uncluster them and reveal more of the detailed information.  

<br>  

```{r leaflet, echo=F, warning=F, message=F}
library(leaflet)
library(leaflet.providers)
library(htmltools)

# Northern Channel Islands
north_ci <- ci_shp %>% 
  filter(poly_id %in% c('San Miguel', 'Santa Rosa', 'Santa Cruz', 'Anacapa'))

north_mpas <- mpa_shp %>% 
  filter(!mpa_id %in% c(7599, 8688, 68808024:68808039, 68808268)) %>% 
  mutate(mpa_type = factor(mpa_type, levels = c("Federal Marine Conservation Area", "Federal Marine Reserve",
                                                "State Marine Reserve", "State Marine Conservation Area")))

# North dive sites
north_sites <- dive_sites %>% 
  filter(island_group == "Northern") %>% 
  rename(x=lon_bin,
         y=lat_bin)

# AIS dives northern CI
north_dives <- ais_dives %>% 
  filter(mpa_definition == "Marine Reserve") %>% 
  # Create a factor from shipname for colors
  mutate(site_frequency = factor(site_frequency, levels = c('high', 'medium', 'low'))) %>% 
  rename(x=lon_bin,
         y=lat_bin)
  

## Colors
pal <- c('slategray3', 'tomato', 'darkcyan', 'firebrick')
mpa_pal <- colorFactor(pal, north_mpas$mpa_type) 

freq <- hcl.colors(n=3, palette = "Heat")
frequency_pal <- colorFactor(freq, north_dives$site_frequency)

# ship_pal <- c(paletteer_d('nationalparkcolors::Arches', 5), paletteer_d('nationalparkcolors::CraterLake', 3), 
#                    "#376597FF", "#E8C533FF", "#979461FF", "#CD5733FF", "#A12A19FF")
# shipname_pal <- colorFactor(ship_pal, north_dives$shipname)

## Map
interactive_map <- leaflet() %>%
  addProviderTiles(providers$Esri.WorldGrayCanvas) %>%
  addPolygons(data=north_mpas,
              smoothFactor = 0.5,
              opacity = 1,
              weight = 1,
              color = "grey60",
              fillColor = NA,
              group = "mpa_outline") %>% 
  addPolygons(data=north_mpas,
              smoothFactor = 0.5,
              opacity = 1,
              weight = 1,
              color = ~mpa_pal(mpa_type),
              fillColor = ~mpa_pal(mpa_type),
              fillOpacity = 0.6,
              group = "mpa_type",
              label = paste("Name:", north_mpas$name)) %>% 
  addCircleMarkers(data=north_sites,
                   lng = north_sites$x,
                   lat = north_sites$y,
                   radius = 5,
                   color = "#8b1a1a",
                   fillColor = "#8b1a1a",
                   stroke = FALSE,
                   fillOpacity = 1,
                   group = "dive_sites",
                   clusterOptions = markerClusterOptions()) %>%
  addCircleMarkers(data=north_dives,
                   lng = north_dives$x,
                   lat = north_dives$y,
                   radius = 5,
                   color = ~frequency_pal(site_frequency),
                   fillColor = ~frequency_pal(site_frequency),
                   stroke = FALSE,
                   fillOpacity = 1,
                   group = "ais_dives",
                   label = paste("Date:", north_dives$date, "<br>",
                                 "Time of Day:", north_dives$time_of_day) %>%
                     lapply(htmltools::HTML),
                   clusterOptions = markerClusterOptions()) %>%
  addLegend("bottomright",
            pal = mpa_pal,
            values = north_mpas$mpa_type,
            title = "MPA Type",
            opacity = 1,
            group = "mpa_type") %>% 
  addLegend("bottomleft",
            pal = frequency_pal,
            values = north_dives$site_frequency,
            title = "Site Frequency",
            opacity = 1,
            group = "ais_dives") %>% 
  addLayersControl(baseGroups = c("mpa_outline"), 
                   overlayGroups = c("mpa_type", "dive_sites", "ais_dives"),
                   options = layersControlOptions(collapsed = FALSE)) %>% 
  hideGroup("mpa_type") %>% 
  hideGroup("dive_sites") %>% 
  hideGroup("ais_dives")

interactive_map
```

## Exploratory {.tabset}

The following are just a few quick exploratory plots on the AIS identified dives. 

### Core Sites

The first plot compares the number of dives at high frequency, medium frequency, and low frequency sites. The second plot looks at the proportion of sites by site frequency within each MPA classification (i.e., proportion of high frequency sites located within MPAs, in MPA buffers, or outside of MPAs). Finally, we compare the number of dives by MPA classification for high frequency dive sites.    

High frequency are sites with the top 25% of visits (aggregated across all vessels and years), sites with more than 3 visits are considered "high frequency" and there are 222 high frequency sites. Medium frequency sites are the next 25% of visits, sites with more than one visit and 3 or less visits are considered "medium" frequency. There are 257 medium frequency sites. The remaining 518 sites, were only visited once and are considered low frequency sites. 

<br>  

```{r core-sites, message=F, warning=F, echo=F}
dives_frequency <- north_dives %>% 
  group_by(year, site_frequency) %>% 
  dplyr::summarize(n_dives = n()) %>% 
  ungroup() %>% 
  group_by(year) %>% 
  mutate(total_annual_dives = sum(n_dives)) %>% 
  ungroup() %>% 
  mutate(prop_dives = n_dives / total_annual_dives)

dive_frequency_plot <- ggplot(data=dives_frequency) + 
  geom_col(aes(x=year, y=n_dives, fill=site_frequency)) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2022,1)) + 
  scale_y_continuous(expand = c(0,0)) + 
  scale_fill_manual(values = freq,
                    labels = c("High", "Medium", "Low")) + 
  labs(x="Year", 
       y="Number of Dives",
       fill="Site Frequency") + 
  theme_bw()

dive_frequency_plot
```
<br> 
58 -76 % of the dives in each year are done at one of the 222 high frequency sites. The proportion of dives at frequency sites are lower in 2020-2022 (58-65%) than previous years (72-76%).  

<br> 

```{r mpa-classification-core-sites, message=F, warning=F, echo=F}
site_mpa_frequency <- ais_dives %>% 
  dplyr::select(mpa_definition, site_id, site_category, site_frequency) %>% 
  distinct() %>% 
  group_by(mpa_definition, site_frequency) %>% 
  mutate(total_sites_in_frequency = n()) %>% 
  ungroup() %>% 
  group_by(mpa_definition, site_frequency, site_category, total_sites_in_frequency) %>% 
  count() %>% 
  mutate(prop_category = n / total_sites_in_frequency)

site_mpa_plot <- ggplot(site_mpa_frequency) + 
  geom_col(aes(x=factor(site_frequency, levels = c('low', 'medium', 'high')), y=prop_category, fill=site_category)) + 
  scale_x_discrete(expand = c(0,0),
                   labels = c("Low", "Medium", "High")) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,1)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(x="Site Frequency",
       y="Proportion of Sites",
       fill = "Site Category") + 
  theme_bw() + 
  facet_wrap(~mpa_definition)

site_mpa_plot
```

<br>
While a majority of the dive sites in each site frequency group (High, Medium, Low) are classified as being outside of MPAs, there is a larger proportion of the medium frequency and high frequency sites within MPAs compared to low frequency sites. This trend is more pronounced when looking at the combined Marine Reserve and Conservation Area definition of "MPA", but is still true when looking only at Marine Reserves as well. 
<br> 
```{r dives-mpa-and-frequency, message=F, warning=F, echo=F}
dives_mpa_frequency <- ais_dives %>% 
  filter(site_frequency == "high") %>% 
  group_by(mpa_definition, year) %>% 
  mutate(total_dives_year = n()) %>% 
  ungroup() %>% 
  group_by(mpa_definition, year, site_category, total_dives_year) %>% 
  summarize(n_dives = n()) %>% 
  ungroup() %>% 
  mutate(prop_dives = n_dives / total_dives_year)

dive_mpa_plot <- ggplot(dives_mpa_frequency) + 
  geom_col(aes(x=year, y=prop_dives, fill=site_category)) +
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2022,1)) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,1)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(x="Year",
       y="Proportion of Dives",
       fill = "Site Category") + 
  theme_bw() + 
  facet_wrap(~mpa_definition)

dive_mpa_plot
```
<br> 

Approximately 23% of high frequency sites fall within Marine Reserve areas. We might expect that if all high frequency sites were visited equally then about 23% of the dives in a given year at these high frequency sites would fall within an MPA. Somewhere between 28-36% of the dives in each year across high frequency sites are at sites within a Marine Reserve. When considering MPAs as Marine Reserves and Conservation Areas, ~38% of the high frequency sites are within MPAs and somewhere between 44 and 59% of the dives in each year across high frequency sites are at sites within a Marine Reserve or Conservation Area. This seems like a good signal for a more rigorous statistical evaluation of dive site preference. 

### Time Trends

A quick look at the difference in number of dives by site category (in MPA, in Buffer, Outside) by month to see if there is a signal for changing preference of dive site locations in October - February, but especially October, during lobster season. Data are separated by month, but are aggregated across 2016 - 2022 (i.e., data for October reflect all dives in October between 2016-2022). Note there is no data for December 2022, so December is only reflective of data from 2016-2021.  

<br> 

```{r time-trends, message=F, warning=F, echo=F}
dives_over_time <- ais_dives %>% 
  group_by(mpa_definition, month, site_category) %>% 
  summarize(n_dives = n()) %>% 
  ungroup() %>% 
  mutate(month = factor(month, levels = c(1:12)))

dives_time_plot <- ggplot(dives_over_time) + 
  geom_col(aes(x=month, y=n_dives, fill=site_category)) + 
  scale_x_discrete(expand = c(0,0),
                   labels = c("Jan", "Feb", "March", "April", "May", "June",
                              "July", "Aug", "Sept", "Oct", "Nov", "Dec"),
                   ) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,600),
                     breaks = seq(0,600,200)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(x = "Month", 
       y = "Number of Dives",
       fill = "Site Category") + 
  theme_bw() + 
  theme(axis.text.x = element_text(angle=90, vjust=0.5)) + 
  facet_wrap(~mpa_definition)

dives_time_plot
```

<br> 
It seems like July - October are the months with the highest amount of diving activity. It also looks like the fraction of dives outside of MPAs is a little bit higher in October than other months with similar amounts of activity (e.g., July - September). 


## Old
### Dives by Year

The total number of dives by year aggregated across all vessels is shown in the first plot and the number of dives by year for each vessel is shown in the second plot. We see an (expected) decline in dive activity from 2019-2020. This plot is reflective of dives from only those vessels who have responded to the survey so far: Spectre, Peace, Raptor, and Vision.         

```{r count-dives, include=F}
ais_positions <- read_csv(file.path(project_data_path, "processed", "ais_stationary_day_positions_2016_2021.csv"))

# Filter positions where hours > 1
# Join to dive site locations
ais_dives <- ais_positions %>% 
  filter(total_hours >= 1) %>% 
  inner_join(dive_sites, by = c("lat_bin", "lon_bin")) %>% 
  left_join(vessel_info %>% 
              dplyr::select(ssvid, shipname),
            by = "ssvid") %>% 
  filter(shipname %in% c("Spectre", "Peace", "Raptor", "Vision"))

## Join to MPA info 
# Join site locations to MPA category
ais_mpa_dives <- ais_dives %>% 
  left_join(all_site_categories, by = "site_id") 
```

```{r dives-year}
# Dives per year (not by MPA)
dives_per_year <- ais_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(year) %>% 
  summarize(n_dives = n()) %>% 
  ungroup()

ggplot(dives_per_year) + 
  geom_col(aes(x=year, y=n_dives),
           fill = 'slategrey',
           alpha = 0.95) + 
  scale_x_continuous(expand = c(0,0)) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,1200),
                     breaks = seq(0,1200,300)) +
  labs(y = 'Number of Dives (all vessels)',
       x = 'Year') + 
  theme_bw()
```

<br>  

For now, we're just looking at the four vessels who have responded to the survey: Spectre, Peace, Raptor, and Vision. 

```{r dives-vessel-year, echo=F}
# Dives per vessel per year (not by MPA)
dives_per_vessel_year <- ais_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(ssvid, shipname, year) %>% 
  summarize(n_dives = n()) %>% 
  ungroup()

ggplot(dives_per_vessel_year) +
  geom_point(aes(x=year, y=n_dives, color=shipname)) + 
  geom_line(aes(x=year, y=n_dives, color=shipname, group=shipname)) + 
  scale_color_manual(values = paletteer::paletteer_d("nationalparkcolors::ArcticGates", n=4)) + 
  scale_x_continuous(expand = c(0.01,0.01)) + 
  scale_y_continuous(expand = c(0.01,0),
                     limits = c(1,400),
                     breaks = seq(0,400,100)) + 
  labs(x = "Year", 
       y = "Number of Dives",
       color = "Shipname") + 
  theme_bw()
```

### Sites by Year

The number of unique dive sites visited each year on the whole was increasing slightly each year, with a noticeable decline from 2019-2020 and continues to increase from 2020 to 2021. These numbers are reflective of only those vessels who have responded to the survey so far: Spectre, Peace, Raptor, and Vision.   
<br>  

```{r unique-sites-per-year, echo=F}
# Unique sites visited per year 
sites_per_year <- ais_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  dplyr::select(year, site_id) %>% 
  distinct() %>% 
  group_by(year) %>% 
  summarize(n_unique_sites = n()) %>% 
  ungroup() 

ggplot(sites_per_year) + 
  geom_col(aes(x=year, y=n_unique_sites),
           fill = 'slategrey',
           alpha = 0.95) + 
  scale_x_continuous(expand = c(0,0)) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,570),
                     breaks = seq(0,550,50)) +
  labs(y = 'Number of Unique Dive Sites',
       x = 'Year') + 
  theme_bw()
```
<br> 

The number of unique sites visited by each vessels seems to be generally in line with survey results, although the Vision reported visiting as many as 500 dive sites per year and that doesn't seem to be reflected in these numbers  
<br>  
```{r unique-sites-per-vessel-year, echo=F}
# Unique sites visited by each vessel in a year 
sites_per_vessel_year <- ais_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  dplyr::select(ssvid, shipname, year, site_id) %>% 
  distinct() %>% 
  group_by(ssvid, shipname, year) %>% 
  summarize(n_unique_sites = n()) %>% 
  ungroup()

ggplot(sites_per_vessel_year) +
  geom_point(aes(x=year, y=n_unique_sites, color=shipname)) + 
  geom_line(aes(x=year, y=n_unique_sites, color=shipname, group=shipname)) + 
  scale_color_manual(values = paletteer::paletteer_d("nationalparkcolors::ArcticGates", n=4)) + 
  scale_x_continuous(expand = c(0.01,0.01)) + 
  scale_y_continuous(expand = c(0.01,0),
                     limits = c(0,275),
                     breaks = seq(0,250,50)) + 
  labs(x = "Year", 
       y = "Number of Unique Sites Visited",
       color = "Shipname") + 
  theme_bw()
```
<br>  
It looks like vessels are visiting anywhere from 60 to 260 unique dive sites every year. This method for classifying dive sites used a less strict categorization, so here's a rough breakdown of how frequently dive sites were visited. 

```{r site-freqency, echo=F}
unique_sites <- ais_dives %>% 
  dplyr::select(lat_bin, lon_bin) %>% 
  distinct() #1399 unique sites

site_freqency_vessel_year <- ais_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(ssvid, shipname, year, site_id) %>% 
  summarize(n_visits_to_site = n()) %>% 
  ungroup()

ggplot(site_freqency_vessel_year) + 
  geom_histogram(aes(x=n_visits_to_site), 
                 fill = 'slategrey',
                 binwidth=3, 
                 boundary = 0) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(0,33,3)) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,1000),
                     breaks = seq(0,1000,250)) + 
  labs(x = "Number of Dives per Site", 
       y = "Freqency") + 
  theme_bw() + 
  facet_wrap(~shipname)
```

Many of the vessels seem to visit the sites only 1-3 times per year. The fraction of sites visited exactly once per year for each vessel is shown in the plot below. The Peace and Vision in particular seem to visit sites only once per year. For the most part since the number of unique sites the vessels are visiting seems to line up pretty well with survey responses, this might not necessarily be wrong but we may want to consider an analysis that differentiates between sites visited once and sites visited repeatedly.     

```{r single-visit, echo=F}
one_visit <- site_freqency_vessel_year %>% 
  group_by(ssvid, shipname, year) %>% 
  mutate(total_dives = sum(n_visits_to_site)) %>% 
  ungroup() %>% 
  filter(n_visits_to_site == 1) %>% 
  group_by(ssvid, shipname, year, total_dives) %>% 
  summarize(n_dives_one_visit = n()) %>% 
  ungroup() %>% 
  mutate(frac_single_visit = n_dives_one_visit / total_dives)

ggplot(one_visit) + 
  geom_col(aes(x=year, y=frac_single_visit, fill=shipname),
           position = 'dodge') + 
  scale_fill_manual(values = paletteer::paletteer_d("nationalparkcolors::ArcticGates", n=4)) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2021,1)) + 
  scale_y_continuous(expand = c(0,0),
                     limits = c(0,1)) +
  labs(y = 'Proportion of Dives',
       x = 'Year',
       fill = 'Shipname') + 
  theme_bw()
```

A breakdown of those sites visited only once per year by site category is shown below.  

```{r single-visit-mpa-category, echo=F}
one_visit_mpa <- ais_mpa_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(ssvid, shipname, year, site_id, site_category) %>% 
  summarize(n_visits_to_site = n()) %>% 
  ungroup() %>% 
  filter(n_visits_to_site == 1) %>% 
  group_by(ssvid, shipname, year) %>% 
  mutate(n_sites_visited_once = n()) %>% 
  ungroup() %>% 
  group_by(ssvid, shipname, year, n_sites_visited_once, site_category) %>% 
  summarize(n_sites = n()) %>% 
  ungroup() %>% 
  mutate(prop_by_category = n_sites / n_sites_visited_once)

ggplot(one_visit_mpa) + 
  geom_col(aes(x=year, y=prop_by_category, fill=site_category)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2021,1)) + 
  scale_y_continuous(expand = c(0,0)) +
  labs(y = 'Proportion of Dives to a Site Once per Year',
       x = 'Year',
       fill = 'Site Category') + 
  theme_bw() + 
  facet_wrap(~shipname)
```

Lastly, let's check the relative breakdown by site category of the dive sites that are visited more than once per year. 

```{r repeat-visit-mpa-category}
repeat_visit_mpa <- ais_mpa_dives %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(ssvid, shipname, year, site_id, site_category) %>% 
  summarize(n_visits_to_site = n()) %>% 
  ungroup() %>% 
  filter(n_visits_to_site > 1) %>% 
  group_by(ssvid, shipname, year) %>% 
  mutate(n_sites_repeated = n()) %>% 
  ungroup() %>% 
  group_by(ssvid, shipname, year, n_sites_repeated, site_category) %>% 
  summarize(n_sites = n()) %>% 
  ungroup() %>% 
  mutate(prop_by_category = n_sites / n_sites_repeated)

ggplot(repeat_visit_mpa) + 
  geom_col(aes(x=year, y=prop_by_category, fill=site_category)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2021,1)) + 
  scale_y_continuous(expand = c(0,0)) +
  labs(y = 'Proportion of Dives to a Site Once per Year',
       x = 'Year',
       fill = 'Site Category') + 
  theme_bw() + 
  facet_wrap(~shipname)
```

## Compare Dives {.tabset}

### Aggregate 

The first plot shows the proportion of total dives by all vessels within each of the site classifications. Again, this is just looking at the four vessels which have responded to the survey so far. On the whole, this indicates that roughly twice as many dives happen outside of MPAs than inside MPAs and MPA buffers. But there are also many more dive sites identified outside of MPAs.  

```{r overall-dives-by-category, echo=F}
# Overall fraction of dives by category 
dive_summary <- ais_mpa_dives %>% 
  dplyr::select(ssvid, date, site_id, site_category) %>% 
  distinct() %>% # Remove any dives classified as being at two MPAs 
  group_by(site_category) %>%
  summarize(n_total_dives = n()) %>% 
  ungroup()

annual_dive_summary <- ais_mpa_dives %>% 
  dplyr::select(ssvid, date, site_id, site_category) %>% 
  mutate(year = lubridate::year(date)) %>% 
  group_by(year) %>% 
  mutate(n_total_dives = n()) %>% 
  ungroup() %>% 
  group_by(year, n_total_dives, site_category) %>%
  summarize(n_dives = n()) %>% 
  ungroup() %>% 
  mutate(frac_dives = n_dives / n_total_dives)

ggplot(annual_dive_summary) + 
  geom_col(aes(x=year, y=frac_dives, fill=site_category)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  scale_x_continuous(expand = c(0,0),
                     breaks = seq(2016,2021,1)) + 
  scale_y_continuous(expand = c(0,0)) +
  labs(y = 'Proportion of Dives',
       x = 'Year',
       fill = 'Dive Site\nCategory') + 
  theme_bw()
```

### By Vessel

The next three graphs show the breakdown of dives by site category for each vessel. The first two are aggregated as total dives from 2016-2021 displayed both as a relative proportion and as an absolute number. Two of the vessels seem to dive about 50/50 inside and outside MPAs while the other two are predominantly diving outside of MPAs. 

The final graph is the breakdown of sites by category for each vessel in each year. 

```{r prop-vessel-dives-by-category, echo=F}
# Fraction of dives by category and vessel
dives_by_category <- ais_mpa_dives %>% 
  dplyr::select(ssvid, shipname, date, site_id, site_category) %>% 
  distinct() %>% # Remove any dives classified as being at two MPAs 
  group_by(ssvid, shipname) %>% 
  mutate(n_total_dives = n()) %>% 
  ungroup() %>% 
  group_by(ssvid, shipname, n_total_dives, site_category) %>%
  summarize(n_dives = n()) %>% 
  ungroup() %>% 
  mutate(frac_of_total = n_dives / n_total_dives)

ggplot(dives_by_category) + 
  geom_col(aes(x=frac_of_total, y=fct_rev(factor(shipname)), fill = site_category)) +
  scale_y_discrete(expand = c(0,0)) + 
  scale_x_continuous(expand = c(0,0)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(y = "Shipname",
       x = "Proportion of Total Dives 2016-2021",
       fill = "Dive Site\nCategory") + 
  theme_bw()
```
<br> 
```{r vessel-dives-category, echo=F}
ggplot(dives_by_category) + 
  geom_col(aes(x=n_dives, y=fct_rev(factor(shipname)), fill = site_category)) +
  scale_y_discrete(expand = c(0,0)) + 
  scale_x_continuous(expand = c(0,0),
                     limits = c(0,1800),
                     breaks = seq(0,1800,600)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(y = "Shipname",
       x = "Total Dives 2016-2021",
       fill = "Dive Site\nCategory") + 
  theme_bw()
```
<br>  
```{r dives-category-overtime, echo=F}
# Dives by category per year
annual_dives_by_category <- ais_mpa_dives %>% 
  dplyr::select(ssvid, shipname, date, site_id, site_category) %>% 
  distinct() %>% # Remove any dives classified as being at two MPAs 
  mutate(year = lubridate::year(date)) %>% 
  group_by(ssvid, shipname, year) %>% 
  mutate(n_annual_dives = n()) %>% 
  ungroup() %>% 
  group_by(ssvid, shipname, year, n_annual_dives, site_category) %>%
  summarize(n_dives = n()) %>% 
  ungroup() %>% 
  mutate(frac_of_total = n_dives / n_annual_dives) 

ggplot(annual_dives_by_category) + 
  geom_col(aes(x=frac_of_total, y=fct_rev(factor(shipname)), fill = site_category)) +
  scale_y_discrete(expand = c(0,0)) + 
  scale_x_continuous(expand = c(0,0)) + 
  scale_fill_manual(values = c("#F4E7C5FF", "#979461FF", "#CD5733FF"),
                    labels = c("In Buffer", "In MPA", "Outside MPA")) + 
  labs(y = "Shipname",
       x = "Proportion of Total Dives 2016-2021",
       fill = "Dive Site\nCategory") + 
  theme_bw() + 
  facet_wrap(~year)
```
